requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pytest
pytest-asyncio
openai>=1.0
